#50 Update tests structure and fix some issues
