+++ Aquality.Selenium/tests/Aquality.Selenium.Tests/Integration/TestApp/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace Aquality.Selenium.Tests.Integration.TestApp
+{
+    internal static class Constants
+    {
+        public const string UrlTheInternet = "http://the-internet.herokuapp.com/";
+        public const string UrlAutomationPractice = "http://automationpractice.com";
+    }
+}
