+++ Aquality.Selenium/tests/Aquality.Selenium.Tests/Integration/TestApp/TheInternet/Forms/AvailableExample.cs
@@ -1,0 +1,22 @@
+﻿using System.ComponentModel;
+
+namespace Aquality.Selenium.Tests.Integration.TestApp.TheInternet.Forms
+{
+    internal enum AvailableExample
+    {
+        [Description("checkboxes")]
+        Checkboxes,
+        [Description("dropdown")]
+        Dropdown,
+        [Description("hovers")]
+        Hovers,
+        [Description("key_presses")]
+        KeyPresses,
+        [Description("infinite_scroll")]
+        InfiniteScroll,
+        [Description("add_remove_elements")]
+        AddRemoveElements,
+        [Description("context_menu")]
+        ContextMenu
+    }
+}
