+++ Aquality.Selenium/tests/Aquality.Selenium.Tests/Integration/TestApp/TheInternet/TheInternetPage.cs
@@ -1,0 +1,10 @@
+﻿namespace Aquality.Selenium.Tests.Integration.TestApp
+{
+    internal static class TheInternetPage
+    {
+        public static string Checkboxes => Constants.UrlTheInternet + "checkboxes";
+        public static string Download => Constants.UrlTheInternet + "download";
+        public static string Dropdown => Constants.UrlTheInternet + "dynamic_controls";
+        public static string Login => Constants.UrlTheInternet + "login";
+    }
+}
