--- Aquality.Selenium/tests/Aquality.Selenium.Tests/TestApp/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace Aquality.Selenium.Tests.TestApp
-{
-    internal static class Constants
-    {
-        public const string UrlTheInternet = "http://the-internet.herokuapp.com/";
-        public const string UrlAutomationPractice = "http://automationpractice.com";
-    }
-}
