--- Aquality.Selenium/tests/Aquality.Selenium.Tests/TestApp/TheInternet/TheInternetPage.cs
@@ -1,10 +1,0 @@
-﻿namespace Aquality.Selenium.Tests.TestApp
-{
-    internal class TheInternetPage
-    {
-        public static string Checkboxes => Constants.UrlTheInternet + "checkboxes";
-        public static string Download => Constants.UrlTheInternet + "download";
-        public static string Dropdown => Constants.UrlTheInternet + "dynamic_controls";
-        public static string Login => Constants.UrlTheInternet + "login";
-    }
-}
